requirements.txt
annotated-types==0.6.0 ; python_version >= "3.9" and python_version < "4.0"
cachetools==5.3.2 ; python_version >= "3.9" and python_version < "4.0"
certifi==2023.11.17 ; python_version >= "3.9" and python_version < "4.0"
charset-normalizer==3.3.2 ; python_version >= "3.9" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.9" and python_version < "4.0" and sys_platform == "win32"
google-api-core==2.14.0 ; python_version >= "3.9" and python_version < "4.0"
google-auth==2.23.4 ; python_version >= "3.9" and python_version < "4.0"
google-cloud-core==2.3.3 ; python_version >= "3.9" and python_version < "4.0"
google-cloud-storage==2.13.0 ; python_version >= "3.9" and python_version < "4.0"
google-crc32c==1.5.0 ; python_version >= "3.9" and python_version < "4.0"
google-resumable-media==2.6.0 ; python_version >= "3.9" and python_version < "4.0"
googleapis-common-protos==1.61.0 ; python_version >= "3.9" and python_version < "4.0"
idna==3.4 ; python_version >= "3.9" and python_version < "4.0"
joblib==1.3.2 ; python_version >= "3.9" and python_version < "4.0"
loguru==0.7.2 ; python_version >= "3.9" and python_version < "4.0"
protobuf==4.25.1 ; python_version >= "3.9" and python_version < "4.0"
pyasn1-modules==0.3.0 ; python_version >= "3.9" and python_version < "4.0"
pyasn1==0.5.1 ; python_version >= "3.9" and python_version < "4.0"
pydantic-core==2.14.5 ; python_version >= "3.9" and python_version < "4.0"
pydantic==2.5.2 ; python_version >= "3.9" and python_version < "4.0"
requests==2.31.0 ; python_version >= "3.9" and python_version < "4.0"
rsa==4.9 ; python_version >= "3.9" and python_version < "4"
typing-extensions==4.8.0 ; python_version >= "3.9" and python_version < "4.0"
urllib3==2.1.0 ; python_version >= "3.9" and python_version < "4.0"
win32-setctime==1.1.0 ; python_version >= "3.9" and python_version < "4.0" and sys_platform == "win32"
